@@ -6,27 +6,27 @@ authors = ["ITBH"]
 license = "GPL-3.0-only"
 
 [tool.poetry.dependencies]
-python = "~=3.8"
-Django = "3.2.18"
-psycopg2 = {version = "2.9.5", extras = ["binary"]}
-django-meta = "2.1.0"
-django-webpack-loader = "==1.8.1"
-django-extensions = "3.2.1"
-django-filer = "2.2.4"
+python = "~=3.10"
+Django = "~4.2.3"
+psycopg2 = {version = "2.9.6", extras = ["binary"]}
+django-meta = "~2.2.0"
+django-webpack-loader = "==2.0.1"
+django-extensions = "3.2.3"
+django-filer = "2.2.6"
 django-polymorphic = "3.1.0"
 blessings = "1.7"
 curtsies = "0.4.1"
 bpython = "0.24"
-django-taggit = "2.1.0"
+django-taggit = "3.1.0"
 xlrd = "2.0.1"
 environs = "9.5.0"
 uWSGI = "2.0.21"
 easy-thumbnails = "2.8.5"
-django-ckeditor = "6.5.1"
+django-ckeditor = "6.6.1"
 django-crispy-forms = "2.0"
-django-import-export = "==2.9.0"
+django-import-export = "3.2.0"
 djangorestframework = "3.14.0"
-django-filter = "2.4.0"
+django-filter = "22.1"
 django-rest-polymorphic = "0.1.10"
 XlsxWriter = "1.4.5"
 rules = "==3.3"
@@ -34,17 +34,17 @@ celery = "==5.2.7"
 redis = "==4.5.1"
 pysolr = "==3.9.0"
 django-haystack = "==3.2.1"
-django-select2 = "==8.1.1"
+django-select2 = "8.1.2"
 django-better-admin-arrayfield = "==1.4.2"
 django-storages = "==1.13.2"
-boto3 = "==1.26.90"
+boto3 = "1.27.0"
 django-simple-history = "==3.3.0"
 django-constance = {version = "==2.9.1", extras = ["database"]}
-wagtail = "3.0.3"
+wagtail = "5.0.2"
 PyNaCl = "==1.5.0"
 django-nested-admin = "4.0.2"
 sentry-sdk = "1.16.0"
-wagtailmenus = "3.1.5"
+wagtailmenus = "3.1.8"
 crispy-bootstrap4 = "^2022.1"
 
 [tool.poetry.dev-dependencies]
