@@ -1,8 +1,8 @@
-from django.conf.urls import url
+from django.urls import re_path
 from . import views
 
 urlpatterns = [
-    url(r"^(?P<message>.+).js$", views.shared_session_view, name="share"),
+    re_path(r"^(?P<message>.+).js$", views.shared_session_view, name="share"),
 ]
 
 urls = urlpatterns, "shared_session", "shared_session"
