@@ -1,8 +1,8 @@
 # Generated by Django 2.2.17 on 2020-11-23 15:23
 
 from django.db import migrations
-import wagtail.core.blocks
-import wagtail.core.fields
+import wagtail.blocks
+import wagtail.fields
 import wagtail.embeds.blocks
 import wagtail.images.blocks
 
@@ -16,11 +16,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="blockpage",
             name="body",
-            field=wagtail.core.fields.StreamField(
+            field=wagtail.fields.StreamField(
                 [
                     (
                         "single_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -32,10 +32,10 @@ class Migration(migrations.Migration):
                                     "video",
                                     wagtail.embeds.blocks.EmbedBlock(required=False),
                                 ),
-                                ("headline", wagtail.core.blocks.CharBlock()),
+                                ("headline", wagtail.blocks.CharBlock()),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -57,7 +57,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "background_color",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[("blue", "Blau"), ("white", "Weiß")]
                                     ),
                                 ),
@@ -66,7 +66,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "side_by_side_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -78,11 +78,11 @@ class Migration(migrations.Migration):
                                     "video",
                                     wagtail.embeds.blocks.EmbedBlock(required=False),
                                 ),
-                                ("headline", wagtail.core.blocks.CharBlock()),
-                                ("sub_headline", wagtail.core.blocks.CharBlock()),
+                                ("headline", wagtail.blocks.CharBlock()),
+                                ("sub_headline", wagtail.blocks.CharBlock()),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -104,7 +104,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "layout",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("image_left", "Bild links, Text rechts"),
                                             ("image_right", "Bild rechts, Text links"),
@@ -113,7 +113,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "background_color",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[("blue", "Blau"), ("white", "Weiß")]
                                     ),
                                 ),
@@ -122,12 +122,12 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "multi_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
-                                ("headline", wagtail.core.blocks.CharBlock()),
+                                ("headline", wagtail.blocks.CharBlock()),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -149,25 +149,25 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "background_color",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[("blue", "Blau"), ("white", "Weiß")]
                                     ),
                                 ),
                                 (
                                     "elements",
-                                    wagtail.core.blocks.StreamBlock(
+                                    wagtail.blocks.StreamBlock(
                                         [
                                             (
                                                 "two_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -183,13 +183,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -212,13 +212,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -235,15 +235,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "three_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -259,13 +259,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -288,13 +288,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -311,15 +311,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "iframe_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(),
+                                                            wagtail.blocks.URLBlock(),
                                                         ),
                                                         (
                                                             "height",
-                                                            wagtail.core.blocks.IntegerBlock(),
+                                                            wagtail.blocks.IntegerBlock(),
                                                         ),
                                                     ]
                                                 ),
@@ -332,9 +332,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "richtext",
-                        wagtail.core.blocks.RichTextBlock(
-                            template="blocks/richtext.html"
-                        ),
+                        wagtail.blocks.RichTextBlock(template="blocks/richtext.html"),
                     ),
                 ]
             ),
