@@ -1,8 +1,8 @@
 # Generated by Django 2.2.17 on 2020-12-17 17:35
 
 from django.db import migrations, models
-import wagtail.core.blocks
-import wagtail.core.fields
+import wagtail.blocks
+import wagtail.fields
 import wagtail.embeds.blocks
 import wagtail.images.blocks
 
@@ -31,11 +31,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="blockpage",
             name="body",
-            field=wagtail.core.fields.StreamField(
+            field=wagtail.fields.StreamField(
                 [
                     (
                         "single_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -45,7 +45,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -57,13 +57,13 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -86,7 +86,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -98,11 +98,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -119,7 +119,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "side_by_side_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -129,7 +129,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -141,19 +141,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "sub_headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Sub-Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -175,7 +175,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -187,7 +187,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "layout",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("image_left", "Bild links, Text rechts"),
                                             ("image_right", "Bild rechts, Text links"),
@@ -196,11 +196,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -217,17 +217,17 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "multi_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -249,7 +249,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -261,11 +261,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -279,19 +279,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "elements",
-                                    wagtail.core.blocks.StreamBlock(
+                                    wagtail.blocks.StreamBlock(
                                         [
                                             (
                                                 "two_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -301,7 +301,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -315,13 +315,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -344,7 +344,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -365,13 +365,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -379,7 +379,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456ae80>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456afa0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456ae80>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456afa0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -389,15 +389,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "three_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -407,7 +407,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -421,13 +421,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -450,7 +450,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -471,13 +471,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -485,7 +485,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456adf0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456adc0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed4576100>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456adf0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456adc0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed4576100>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -495,15 +495,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "iframe_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(),
+                                                            wagtail.blocks.URLBlock(),
                                                         ),
                                                         (
                                                             "height",
-                                                            wagtail.core.blocks.IntegerBlock(
+                                                            wagtail.blocks.IntegerBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -518,9 +518,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "richtext",
-                        wagtail.core.blocks.RichTextBlock(
-                            template="blocks/richtext.html"
-                        ),
+                        wagtail.blocks.RichTextBlock(template="blocks/richtext.html"),
                     ),
                 ]
             ),
@@ -528,11 +526,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="frontpage",
             name="body",
-            field=wagtail.core.fields.StreamField(
+            field=wagtail.fields.StreamField(
                 [
                     (
                         "single_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -542,7 +540,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -554,13 +552,13 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -583,7 +581,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -595,11 +593,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -616,7 +614,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "side_by_side_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -626,7 +624,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -638,19 +636,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "sub_headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Sub-Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -672,7 +670,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -684,7 +682,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "layout",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("image_left", "Bild links, Text rechts"),
                                             ("image_right", "Bild rechts, Text links"),
@@ -693,11 +691,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -714,17 +712,17 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "multi_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -746,7 +744,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -758,11 +756,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -776,19 +774,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "elements",
-                                    wagtail.core.blocks.StreamBlock(
+                                    wagtail.blocks.StreamBlock(
                                         [
                                             (
                                                 "two_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -798,7 +796,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -812,13 +810,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -841,7 +839,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -862,13 +860,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -876,7 +874,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456aee0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456af10>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456aee0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456af10>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -886,15 +884,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "three_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -904,7 +902,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -918,13 +916,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -947,7 +945,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -968,13 +966,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -982,7 +980,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456afd0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed456afa0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f9ed4577100>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456afd0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed456afa0>), ('text_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f9ed4577100>), ('text_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -992,15 +990,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "iframe_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(),
+                                                            wagtail.blocks.URLBlock(),
                                                         ),
                                                         (
                                                             "height",
-                                                            wagtail.core.blocks.IntegerBlock(
+                                                            wagtail.blocks.IntegerBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -1015,9 +1013,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "richtext",
-                        wagtail.core.blocks.RichTextBlock(
-                            template="blocks/richtext.html"
-                        ),
+                        wagtail.blocks.RichTextBlock(template="blocks/richtext.html"),
                     ),
                 ],
                 null=True,
