@@ -1,8 +1,8 @@
 # Generated by Django 3.2.15 on 2022-09-29 08:14
 
 from django.db import migrations
-import wagtail.core.blocks
-import wagtail.core.fields
+import wagtail.blocks
+import wagtail.fields
 import wagtail.embeds.blocks
 import wagtail.images.blocks
 
@@ -16,11 +16,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="blockpage",
             name="body",
-            field=wagtail.core.fields.StreamField(
+            field=wagtail.fields.StreamField(
                 [
                     (
                         "single_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -36,7 +36,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "image_breakpoint",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("sm", ">=576px"),
                                             ("md", ">=768px"),
@@ -50,7 +50,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -62,13 +62,13 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -91,7 +91,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -103,16 +103,16 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -129,7 +129,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "side_by_side_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -145,7 +145,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "image_breakpoint",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("sm", ">=576px"),
                                             ("md", ">=768px"),
@@ -159,7 +159,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -171,19 +171,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "sub_headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Sub-Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -205,7 +205,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -217,7 +217,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "layout",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("image_left", "Bild links, Text rechts"),
                                             ("image_right", "Bild rechts, Text links"),
@@ -226,11 +226,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -244,25 +244,25 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                             ]
                         ),
                     ),
                     (
                         "multi_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -285,7 +285,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -297,16 +297,16 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -320,19 +320,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "elements",
-                                    wagtail.core.blocks.StreamBlock(
+                                    wagtail.blocks.StreamBlock(
                                         [
                                             (
                                                 "two_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -349,7 +349,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_breakpoint",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "sm",
@@ -375,7 +375,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -389,13 +389,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -418,7 +418,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -439,7 +439,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -460,13 +460,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -474,7 +474,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2a40>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2b30>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2a40>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2b30>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -484,15 +484,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "three_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -509,7 +509,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_breakpoint",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "sm",
@@ -535,7 +535,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -549,13 +549,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -578,7 +578,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -599,7 +599,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -620,13 +620,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -634,7 +634,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e29e0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2ad0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2c20>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e29e0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2ad0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2c20>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -644,15 +644,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "iframe_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(),
+                                                            wagtail.blocks.URLBlock(),
                                                         ),
                                                         (
                                                             "height",
-                                                            wagtail.core.blocks.IntegerBlock(
+                                                            wagtail.blocks.IntegerBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -661,7 +661,7 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "dll_element_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "image",
@@ -678,7 +678,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "image_breakpoint",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("sm", ">=576px"),
                                                                     ("md", ">=768px"),
@@ -692,7 +692,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "alt_text",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                 label="Alt Text",
                                                                 required=False,
@@ -706,13 +706,13 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "subline",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 required=False
                                                             ),
                                                         ),
                                                         (
                                                             "text",
-                                                            wagtail.core.blocks.RichTextBlock(
+                                                            wagtail.blocks.RichTextBlock(
                                                                 features=[
                                                                     "bold",
                                                                     "italic",
@@ -735,7 +735,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "text_alignment",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("left", "Links"),
                                                                     (
@@ -750,7 +750,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "image_alignment",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("left", "Links"),
                                                                     (
@@ -765,13 +765,13 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "link_text",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 required=False
                                                             ),
                                                         ),
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(
+                                                            wagtail.blocks.URLBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -786,9 +786,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "richtext",
-                        wagtail.core.blocks.RichTextBlock(
-                            template="blocks/richtext.html"
-                        ),
+                        wagtail.blocks.RichTextBlock(template="blocks/richtext.html"),
                     ),
                 ],
                 blank=True,
@@ -797,11 +795,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name="frontpage",
             name="body",
-            field=wagtail.core.fields.StreamField(
+            field=wagtail.fields.StreamField(
                 [
                     (
                         "single_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -817,7 +815,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "image_breakpoint",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("sm", ">=576px"),
                                             ("md", ">=768px"),
@@ -831,7 +829,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -843,13 +841,13 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -872,7 +870,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -884,16 +882,16 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -910,7 +908,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "side_by_side_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "image",
@@ -926,7 +924,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "image_breakpoint",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("sm", ">=576px"),
                                             ("md", ">=768px"),
@@ -940,7 +938,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "alt_text",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                         label="Alt Text",
                                         required=False,
@@ -952,19 +950,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "sub_headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Sub-Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -986,7 +984,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -998,7 +996,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "layout",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("image_left", "Bild links, Text rechts"),
                                             ("image_right", "Bild rechts, Text links"),
@@ -1007,11 +1005,11 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -1025,25 +1023,25 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                             ]
                         ),
                     ),
                     (
                         "multi_element_block",
-                        wagtail.core.blocks.StructBlock(
+                        wagtail.blocks.StructBlock(
                             [
                                 (
                                     "headline",
-                                    wagtail.core.blocks.CharBlock(
+                                    wagtail.blocks.CharBlock(
                                         label="Überschrift", required=False
                                     ),
                                 ),
                                 (
                                     "text",
-                                    wagtail.core.blocks.RichTextBlock(
+                                    wagtail.blocks.RichTextBlock(
                                         features=[
                                             "bold",
                                             "italic",
@@ -1066,7 +1064,7 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "text_alignment",
-                                    wagtail.core.blocks.ChoiceBlock(
+                                    wagtail.blocks.ChoiceBlock(
                                         choices=[
                                             ("left", "Links"),
                                             ("center", "Zentriert"),
@@ -1078,16 +1076,16 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "link_text",
-                                    wagtail.core.blocks.CharBlock(required=False),
+                                    wagtail.blocks.CharBlock(required=False),
                                 ),
-                                ("url", wagtail.core.blocks.URLBlock(required=False)),
+                                ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
-                                    wagtail.core.blocks.StructBlock(
+                                    wagtail.blocks.StructBlock(
                                         [
                                             (
                                                 "background_color",
-                                                wagtail.core.blocks.CharBlock(
+                                                wagtail.blocks.CharBlock(
                                                     help_text="Hex Wert ohne #",
                                                     label="Hintergrund Farbe",
                                                     max_length=6,
@@ -1101,19 +1099,19 @@ class Migration(migrations.Migration):
                                 ),
                                 (
                                     "elements",
-                                    wagtail.core.blocks.StreamBlock(
+                                    wagtail.blocks.StreamBlock(
                                         [
                                             (
                                                 "two_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -1130,7 +1128,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_breakpoint",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "sm",
@@ -1156,7 +1154,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -1170,13 +1168,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -1199,7 +1197,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -1220,7 +1218,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -1241,13 +1239,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -1255,7 +1253,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2aa0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2b90>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2aa0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2b90>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -1265,15 +1263,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "three_column_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "content",
-                                                            wagtail.core.blocks.StreamBlock(
+                                                            wagtail.blocks.StreamBlock(
                                                                 [
                                                                     (
                                                                         "dll_element_block",
-                                                                        wagtail.core.blocks.StructBlock(
+                                                                        wagtail.blocks.StructBlock(
                                                                             [
                                                                                 (
                                                                                     "image",
@@ -1290,7 +1288,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_breakpoint",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "sm",
@@ -1316,7 +1314,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "alt_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                                         label="Alt Text",
                                                                                         required=False,
@@ -1330,13 +1328,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "subline",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "text",
-                                                                                    wagtail.core.blocks.RichTextBlock(
+                                                                                    wagtail.blocks.RichTextBlock(
                                                                                         features=[
                                                                                             "bold",
                                                                                             "italic",
@@ -1359,7 +1357,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "text_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -1380,7 +1378,7 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "image_alignment",
-                                                                                    wagtail.core.blocks.ChoiceBlock(
+                                                                                    wagtail.blocks.ChoiceBlock(
                                                                                         choices=[
                                                                                             (
                                                                                                 "left",
@@ -1401,13 +1399,13 @@ class Migration(migrations.Migration):
                                                                                 ),
                                                                                 (
                                                                                     "link_text",
-                                                                                    wagtail.core.blocks.CharBlock(
+                                                                                    wagtail.blocks.CharBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
                                                                                 (
                                                                                     "url",
-                                                                                    wagtail.core.blocks.URLBlock(
+                                                                                    wagtail.blocks.URLBlock(
                                                                                         required=False
                                                                                     ),
                                                                                 ),
@@ -1415,7 +1413,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2b60>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2b00>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.core.rich_text.RichText object at 0x7f0da74e2bc0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2b60>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2b00>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7f0da74e2bc0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -1425,15 +1423,15 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "iframe_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(),
+                                                            wagtail.blocks.URLBlock(),
                                                         ),
                                                         (
                                                             "height",
-                                                            wagtail.core.blocks.IntegerBlock(
+                                                            wagtail.blocks.IntegerBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -1442,7 +1440,7 @@ class Migration(migrations.Migration):
                                             ),
                                             (
                                                 "dll_element_block",
-                                                wagtail.core.blocks.StructBlock(
+                                                wagtail.blocks.StructBlock(
                                                     [
                                                         (
                                                             "image",
@@ -1459,7 +1457,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "image_breakpoint",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("sm", ">=576px"),
                                                                     ("md", ">=768px"),
@@ -1473,7 +1471,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "alt_text",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 help_text="Überschreibt den Standard Alt Text des Bildes.",
                                                                 label="Alt Text",
                                                                 required=False,
@@ -1487,13 +1485,13 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "subline",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 required=False
                                                             ),
                                                         ),
                                                         (
                                                             "text",
-                                                            wagtail.core.blocks.RichTextBlock(
+                                                            wagtail.blocks.RichTextBlock(
                                                                 features=[
                                                                     "bold",
                                                                     "italic",
@@ -1516,7 +1514,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "text_alignment",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("left", "Links"),
                                                                     (
@@ -1531,7 +1529,7 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "image_alignment",
-                                                            wagtail.core.blocks.ChoiceBlock(
+                                                            wagtail.blocks.ChoiceBlock(
                                                                 choices=[
                                                                     ("left", "Links"),
                                                                     (
@@ -1546,13 +1544,13 @@ class Migration(migrations.Migration):
                                                         ),
                                                         (
                                                             "link_text",
-                                                            wagtail.core.blocks.CharBlock(
+                                                            wagtail.blocks.CharBlock(
                                                                 required=False
                                                             ),
                                                         ),
                                                         (
                                                             "url",
-                                                            wagtail.core.blocks.URLBlock(
+                                                            wagtail.blocks.URLBlock(
                                                                 required=False
                                                             ),
                                                         ),
@@ -1567,9 +1565,7 @@ class Migration(migrations.Migration):
                     ),
                     (
                         "richtext",
-                        wagtail.core.blocks.RichTextBlock(
-                            template="blocks/richtext.html"
-                        ),
+                        wagtail.blocks.RichTextBlock(template="blocks/richtext.html"),
                     ),
                 ],
                 null=True,
