@@ -1,15 +1,15 @@
-# Generated by Django 3.2.15 on 2022-09-29 09:04
+# Generated by Django 3.2.18 on 2023-05-08 09:11
 
 from django.db import migrations
 import wagtail.blocks
-import wagtail.fields
 import wagtail.embeds.blocks
+import wagtail.fields
 import wagtail.images.blocks
 
 
 class Migration(migrations.Migration):
     dependencies = [
-        ("cms", "0006_auto_20220929_1014"),
+        ("cms", "0009_auto_20230207_1315"),
     ]
 
     operations = [
@@ -101,10 +101,7 @@ class Migration(migrations.Migration):
                                         required=False,
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
@@ -242,10 +239,7 @@ class Migration(migrations.Migration):
                                         label="Erweiterte Einstellungen",
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                             ]
                         ),
@@ -295,10 +289,7 @@ class Migration(migrations.Migration):
                                         required=False,
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
@@ -474,7 +465,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a2c0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a3b0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8070>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8160>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -634,7 +625,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a260>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a350>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a4a0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a80a0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8040>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8250>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -790,6 +781,7 @@ class Migration(migrations.Migration):
                     ),
                 ],
                 blank=True,
+                use_json_field=True,
             ),
         ),
         migrations.AlterField(
@@ -880,10 +872,7 @@ class Migration(migrations.Migration):
                                         required=False,
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
@@ -1021,10 +1010,7 @@ class Migration(migrations.Migration):
                                         label="Erweiterte Einstellungen",
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                             ]
                         ),
@@ -1074,10 +1060,7 @@ class Migration(migrations.Migration):
                                         required=False,
                                     ),
                                 ),
-                                (
-                                    "link_text",
-                                    wagtail.blocks.CharBlock(required=False),
-                                ),
+                                ("link_text", wagtail.blocks.CharBlock(required=False)),
                                 ("url", wagtail.blocks.URLBlock(required=False)),
                                 (
                                     "advanced_options",
@@ -1253,7 +1236,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a320>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a410>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8070>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8160>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=2,
                                                                 min_num=2,
                                                             ),
@@ -1413,7 +1396,7 @@ class Migration(migrations.Migration):
                                                                         ),
                                                                     )
                                                                 ],
-                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a3e0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a380>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0x7ff12e58a440>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
+                                                                default="[('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a80a0>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8040>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)])), ('dll_element_block', StructValue([('image', None), ('mobile_image', None), ('image_breakpoint', None), ('alt_text', None), ('video', None), ('subline', None), ('text', <wagtail.rich_text.RichText object at 0xffff786a8250>), ('text_alignment', 'center'), ('image_alignment', 'center'), ('link_text', None), ('url', None)]))]",
                                                                 max_num=3,
                                                                 min_num=3,
                                                             ),
@@ -1569,6 +1552,7 @@ class Migration(migrations.Migration):
                     ),
                 ],
                 null=True,
+                use_json_field=True,
             ),
         ),
     ]
