@@ -1,6 +1,4 @@
 import django.dispatch
 
 
-session_replaced = django.dispatch.Signal(
-    providing_args=["request", "src_domain", "dest_domain", "was_empty"]
-)
+session_replaced = django.dispatch.Signal()
